modified:   README A few works about make. 	modified:   configure.ac Version 20180928
Major change in handling options, and now all options
have an easier to understand long name.
	modified:   dwarfdump/command_options.c
	modified:   dwarfdump/glflags.c
	modified:   dwarfdump/glflags.h

Avoids small warning.
	modified:   libdwarf/dwgetopt.c

The .h it creates now usable with C++
	modified:   libdwarf/gennames.c
updated date
	modified:   libdwarf/libdwarf_version.h

+++ libdwarf/dwgetopt.c
@@ -94,8 +94,8 @@ dwoptnamematches(
 {
 
     const char *eq = 0;
-    unsigned namelen = 0;
-    unsigned arglen = 0;
+    size_t namelen = 0;
+    size_t arglen = 0;
     int d = 0;
 
     for(eq = iplace; *eq; ++eq) {
